pyproject.toml
[project]
name = "nano-llm"
version = "0.1.0"
description = "A Nano-like CLI editor with LLM capabilities"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.3.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
]
